Реализовал компонент трип-контейнера. Бета-версия
